requirements.txt
Django
djangorestframework
django-admin-honeypot
django-admin-honeypot-updated-2021
mysqlclient
python-dotenv
django-maintenance-mode
djangorestframework-simplejwt
djangorestframework-simplejwt[crypto]
#jupyterlab
#pydicom
matplotlib
numpy
django-phonenumber-field
phonenumbers
django-countries
pillow
django-cors-headers